feat(sql): add support for temporal SQL clauses and fix CI
- Implements handling for `FOR SYSTEM_TIME` and `FOR VALID_TIME` clauses in SQL queries.
- Adds `temporal_context` to `Query` and propagates it to the query planner.
- Adds `serialized_size` method to `PropertyValue` and `PropertyMap` to fix benchmark compilation.
- Fixes flaky `test_append_async_mode` by using `wal.shutdown()`.
- Adds comprehensive integration tests in `tests/sql_temporal_integration.rs`.
