+++ .jules/warden.md
@@ -23,27 +23,6 @@ Implemented `tests/warden_compliance.rs` to simulate a DoS scenario with high-co
 **Defense:** Automated Check Script
 Created `scripts/security_audit.sh` to automatically check for and run `cargo-audit` if available. This standardizes the security check process for agents and CI.
 
-## 2026-02-01 - SIMD & Public API Hardening
-
-**Threat:** Buffer Over-read in Unsafe SIMD
-The `unsafe` SIMD functions in `src/core/vector.rs` rely on the caller to ensure input slices have equal lengths. If this contract is violated, pointer arithmetic could read past the end of the second buffer (Undefined Behavior).
-
-**Defense:** Assertion Hardening
-Added `debug_assert_eq!(a.len(), b.len())` to internal safe wrapper functions in `src/core/vector.rs` (`dot_and_magnitudes`, `squared_diff_sum`, `dot_product_sum`). This enforces the safety contract before entering `unsafe` SIMD blocks (`dot_and_magnitudes_avx2`, etc.), ensuring validation runs on all hardware (including CI runners without AVX) while protecting the unsafe inner functions.
-
-**Threat:** Resource Exhaustion via Public API
-Public HTTP and MCP endpoints could be vectors for DoS attacks via large inputs (e.g., massive vectors, deep traversals).
-
-**Verification:** Audit & Stress Test
-- **Verified:** `WalRingBuffer` backpressure mechanism passed high-contention stress test (`tests/warden_compliance.rs`).
-- **Audited:** `src/mcp/server.rs` and `src/http/handlers.rs`. Confirmed that resource limits are strictly enforced:
-  - `MAX_TRAVERSAL_DEPTH` (10)
-  - `MAX_RESULT_LIMIT` (10,000)
-  - `MAX_VECTOR_K` (1,000)
-  - Vector dimensions validated against index configuration.
-  - WAL segment reading limits file size to 1GB to prevent memory map exhaustion.
-
 ## Open Risks
 - `usearch` FFI boundaries rely on the C++ library behaving correctly regarding pointer validity. We added panic guards for null pointers, but full memory safety depends on `usearch` correctness.
 - The `usearch` dependency points to a fork (`madmax983/USearch`). This fork contains Rust-specific fixes (move semantics) not yet in upstream. We have pinned the specific commit to ensure stability, but future upstream security patches will need manual cherry-picking.
-- `mmap` usage in `src/storage/wal/segment_reader.rs` is inherently unsafe against external file truncation (SIGBUS risk), though file size is checked before mapping.
