+++ src/index/vector/temporal.rs
@@ -72,7 +72,7 @@
 //! - `docs/adr/0017-temporal-vector-strategy.md` - Design decisions
 //! - `docs/VECTOR_SEARCH_DESIGN.md` - Overall vector search architecture
 
-use std::collections::{HashMap, HashSet};
+use std::collections::{BTreeMap, HashMap, HashSet};
 use std::sync::Arc;
 use std::time::Duration;
 
@@ -83,32 +83,870 @@ use crate::core::id::NodeId;
 use crate::core::temporal::{TimeRange, Timestamp};
 use crate::core::vector::{cosine_similarity, euclidean_distance};
 use crate::index::vector::hnsw::HnswIndex;
-use crate::index::vector::{DistanceMetric, TemporalSearchResults, VectorIndex};
+use crate::index::vector::{DistanceMetric, HnswConfig, TemporalSearchResults, VectorIndex};
 use crate::utils::{Error, Result, TemporalError, VectorError};
 
 #[cfg(feature = "observability")]
 use tracing;
 
-// Submodules
-/// Configuration types for temporal vector indexing.
-pub mod config;
-/// Observers for reacting to storage events.
-pub mod observer;
-pub(crate) mod snapshot;
-/// Statistics and monitoring types.
-pub mod stats;
+/// Retention policy for snapshot cleanup.
+///
+/// Determines which snapshots to keep and which to prune.
+#[derive(Debug, Clone, PartialEq)]
+pub enum RetentionPolicy {
+    /// Keep all snapshots (no automatic pruning).
+    KeepAll,
 
-// Re-exports
-pub use config::{DriftMetric, RetentionPolicy, SnapshotStrategy, TemporalVectorConfig};
-pub use observer::VectorIndexObserver;
-pub use stats::{MemoryStats, SnapshotInfo};
+    /// Keep only the most recent N snapshots.
+    KeepN(usize),
 
-// Internal imports
-use config::{MAX_ACCUMULATED_CHANGES, MAX_SNAPSHOT_RETRIES};
-use snapshot::{
-    DeltaIndex, SnapshotData, SnapshotIndex, SnapshotMetadata, VectorSnapshot, VectorState,
-};
+    /// Keep snapshots within a time duration from now.
+    KeepDuration(Duration),
+}
 
+impl Default for RetentionPolicy {
+    fn default() -> Self {
+        RetentionPolicy::KeepN(100)
+    }
+}
+
+/// Configuration for temporal vector indexing.
+///
+/// This struct encapsulates all parameters needed to configure temporal vector
+/// indexing with HNSW snapshots.
+///
+/// # Examples
+///
+/// ```rust
+/// use gallifreydb::index::vector::temporal::{TemporalVectorConfig, SnapshotStrategy, RetentionPolicy};
+/// use gallifreydb::index::vector::{HnswConfig, DistanceMetric};
+///
+/// // Default configuration (transaction-based, every 10 transactions)
+/// let hnsw_config = HnswConfig::new(384, DistanceMetric::Cosine);
+/// let config = TemporalVectorConfig::default_with_hnsw(hnsw_config);
+///
+/// // Custom hybrid strategy with retention policy
+/// let config = TemporalVectorConfig {
+///     snapshot_strategy: SnapshotStrategy::Hybrid {
+///         transaction_interval: 10,
+///         time_interval_secs: 3600,  // Hourly
+///         change_threshold: 0.1,      // 10% changed
+///     },
+///     retention_policy: RetentionPolicy::KeepN(50),
+///     max_snapshots: 100,
+///     full_snapshot_interval: 10,
+///     hnsw_config: Some(HnswConfig::new(384, DistanceMetric::Cosine)),
+/// };
+/// ```
+/// Maximum number of retries when creating a snapshot due to races (default: 3)
+const MAX_SNAPSHOT_RETRIES: usize = 3;
+
+/// Maximum depth of delta chain traversal before forcing full snapshot (default: 10)
+/// This prevents unbounded delta chains and ensures reconstruction performance.
+/// Mirrors the full_snapshot_interval default value.
+const MAX_DELTA_CHAIN_DEPTH: usize = 10;
+
+/// Minimum capacity estimate for HashMap pre-allocation (default: 100)
+/// Used when estimating capacity for vector collections to avoid excessive resizing.
+///
+/// **Justification**: 100 is a reasonable baseline that:
+/// - Avoids excessive allocations for small datasets (most vectors fit in 1 allocation)
+/// - Prevents too many resizes for medium datasets
+/// - Has negligible overhead (~800 bytes for empty capacity-100 HashMap)
+/// - Aligns with common batch sizes in vector databases
+const MIN_CAPACITY_ESTIMATE: usize = 100;
+
+/// Maximum accumulated changes before forcing a full snapshot (default: 100,000)
+///
+/// If `changes_accumulated` exceeds this threshold, force a full snapshot
+/// regardless of `full_snapshot_interval`. This prevents unbounded memory growth
+/// in write-heavy workloads.
+///
+/// **Justification**: 100k NodeIds × 8 bytes = 800 KB overhead, which is acceptable
+/// but starting to impact memory. Forcing a full snapshot resets the accumulator.
+const MAX_ACCUMULATED_CHANGES: usize = 100_000;
+
+/// Configuration for temporal vector index with snapshot management.
+///
+/// Controls how and when snapshots are created, retained, and pruned to balance
+/// memory usage, query performance, and historical depth.
+#[derive(Debug, Clone, PartialEq)]
+pub struct TemporalVectorConfig {
+    /// Snapshot creation strategy
+    pub snapshot_strategy: SnapshotStrategy,
+
+    /// Snapshot retention policy (default: KeepN(100))
+    pub retention_policy: RetentionPolicy,
+
+    /// Maximum number of snapshots to retain (default: 100)
+    ///
+    /// When this limit is exceeded, the oldest snapshots are removed.
+    /// This prevents unbounded storage growth.
+    pub max_snapshots: usize,
+
+    /// Interval for creating full snapshots (default: 10)
+    ///
+    /// After this many delta snapshots, a new full snapshot is created.
+    /// Higher values save memory but increase reconstruction time.
+    /// Lower values increase memory but improve query speed.
+    pub full_snapshot_interval: usize,
+
+    /// Base HNSW configuration for all indexes (current + snapshots)
+    ///
+    /// If `None`, the temporal index will use the existing vector index's configuration.
+    /// This allows enabling temporal features on an already-configured vector index.
+    pub hnsw_config: Option<HnswConfig>,
+}
+
+impl TemporalVectorConfig {
+    /// Validates the configuration.
+    ///
+    /// # Errors
+    /// Returns an error if:
+    /// - `full_snapshot_interval` exceeds `MAX_DELTA_CHAIN_DEPTH`
+    /// - `max_snapshots` is 0
+    ///
+    /// # Safety
+    /// This validation prevents delta chain depth from exceeding the hard limit,
+    /// which would cause get_vector() and to_hashmap() to return partial/empty results.
+    pub fn validate(&self) -> Result<()> {
+        if self.full_snapshot_interval > MAX_DELTA_CHAIN_DEPTH {
+            return Err(VectorError::IndexError(format!(
+                "full_snapshot_interval ({}) exceeds MAX_DELTA_CHAIN_DEPTH ({}). \
+                     This would cause delta chains to exceed traversal depth limits, \
+                     leading to silent data loss. Reduce full_snapshot_interval to at most {}.",
+                self.full_snapshot_interval, MAX_DELTA_CHAIN_DEPTH, MAX_DELTA_CHAIN_DEPTH
+            ))
+            .into());
+        }
+
+        if self.max_snapshots == 0 {
+            return Err(
+                VectorError::IndexError("max_snapshots must be at least 1".to_string()).into(),
+            );
+        }
+
+        Ok(())
+    }
+
+    /// Creates a default configuration with the given HNSW config.
+    ///
+    /// Defaults:
+    /// - Strategy: TransactionInterval(10) - mirrors anchor+delta pattern
+    /// - Retention: KeepN(100)
+    /// - Max snapshots: 100
+    /// - Full snapshot interval: 10
+    pub fn default_with_hnsw(hnsw_config: HnswConfig) -> Self {
+        TemporalVectorConfig {
+            snapshot_strategy: SnapshotStrategy::TransactionInterval(10),
+            retention_policy: RetentionPolicy::KeepN(100),
+            max_snapshots: 100,
+            full_snapshot_interval: 10,
+            hnsw_config: Some(hnsw_config),
+        }
+    }
+
+    /// Creates a default temporal-only configuration without HNSW config.
+    ///
+    /// Use this when a vector index already exists and you only want to enable
+    /// temporal features. The existing vector index's HNSW configuration will be used.
+    ///
+    /// Defaults:
+    /// - Strategy: TransactionInterval(10) - mirrors anchor+delta pattern
+    /// - Retention: KeepN(100)
+    /// - Max snapshots: 100
+    /// - Full snapshot interval: 10
+    /// - hnsw_config: None (use existing)
+    pub fn default_temporal_only() -> Self {
+        TemporalVectorConfig {
+            snapshot_strategy: SnapshotStrategy::TransactionInterval(10),
+            retention_policy: RetentionPolicy::KeepN(100),
+            max_snapshots: 100,
+            full_snapshot_interval: 10,
+            hnsw_config: None,
+        }
+    }
+
+    /// Creates a configuration for time-based snapshots.
+    pub fn with_time_interval(hnsw_config: HnswConfig, interval_secs: u64) -> Self {
+        TemporalVectorConfig {
+            snapshot_strategy: SnapshotStrategy::TimeInterval(interval_secs),
+            retention_policy: RetentionPolicy::KeepN(100),
+            max_snapshots: 100,
+            full_snapshot_interval: 10,
+            hnsw_config: Some(hnsw_config),
+        }
+    }
+
+    /// Creates a configuration for change-based snapshots.
+    pub fn with_change_threshold(hnsw_config: HnswConfig, threshold: f64) -> Self {
+        TemporalVectorConfig {
+            snapshot_strategy: SnapshotStrategy::ChangeThreshold(threshold),
+            retention_policy: RetentionPolicy::KeepN(100),
+            max_snapshots: 100,
+            full_snapshot_interval: 10,
+            hnsw_config: Some(hnsw_config),
+        }
+    }
+}
+
+/// Snapshot creation strategies.
+///
+/// Determines when temporal vector index snapshots are created.
+///
+/// # Trade-offs
+///
+/// | Strategy | Pros | Cons |
+/// |----------|------|------|
+/// | TransactionInterval | Predictable snapshot count | May miss time-based patterns |
+/// | TimeInterval | Captures time-based changes | Uneven snapshot distribution |
+/// | ChangeThreshold | Adaptive to workload | Unpredictable snapshot count |
+/// | Hybrid | Combines benefits | More complex configuration |
+///
+/// # Examples
+///
+/// ```rust
+/// use gallifreydb::index::vector::temporal::SnapshotStrategy;
+///
+/// // Transaction-based (default): snapshot every 10 transactions
+/// let strategy = SnapshotStrategy::TransactionInterval(10);
+///
+/// // Time-based: snapshot every hour
+/// let strategy = SnapshotStrategy::TimeInterval(3600);
+///
+/// // Change-based: snapshot when 10% of vectors changed
+/// let strategy = SnapshotStrategy::ChangeThreshold(0.1);
+///
+/// // Hybrid: whichever triggers first
+/// let strategy = SnapshotStrategy::Hybrid {
+///     transaction_interval: 10,
+///     time_interval_secs: 3600,
+///     change_threshold: 0.1,
+/// };
+/// ```
+#[derive(Debug, Clone, PartialEq)]
+pub enum SnapshotStrategy {
+    /// Create snapshot every N write transactions.
+    ///
+    /// Mirrors anchor+delta pattern (default: 10).
+    /// Provides predictable snapshot frequency regardless of time.
+    TransactionInterval(usize),
+
+    /// Create snapshot at fixed time intervals (seconds).
+    ///
+    /// Example: 3600 for hourly snapshots.
+    /// Good for time-series analysis of semantic drift.
+    TimeInterval(u64),
+
+    /// Create snapshot when significant changes occur.
+    ///
+    /// Threshold is fraction of total vectors changed (0.0-1.0).
+    /// Example: 0.1 means snapshot when 10% of vectors change.
+    /// Adaptive to workload intensity.
+    ChangeThreshold(f64),
+
+    /// Hybrid: use whichever trigger fires first.
+    ///
+    /// Combines benefits of all strategies.
+    /// Ensures snapshots on any significant event.
+    Hybrid {
+        /// Transaction interval threshold
+        transaction_interval: usize,
+        /// Time interval in seconds
+        time_interval_secs: u64,
+        /// Change threshold (0.0-1.0)
+        change_threshold: f64,
+    },
+}
+
+/// Metric for measuring semantic drift between vector embeddings.
+///
+/// Different metrics capture different aspects of how meaning has changed:
+/// - **Cosine**: Angular difference (independent of magnitude)
+/// - **Euclidean**: Spatial distance (sensitive to magnitude)
+/// - **Angular**: Actual geometric angle in radians
+///
+/// # Examples
+///
+/// ```rust
+/// use gallifreydb::index::vector::temporal::DriftMetric;
+///
+/// let metric = DriftMetric::default(); // Cosine
+/// ```
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+pub enum DriftMetric {
+    /// Cosine distance: 1.0 - cosine_similarity
+    ///
+    /// Range: [0, 2] for normalized vectors, typically [0, 1]
+    /// Most interpretable for semantic embeddings.
+    /// Value of 0 = identical meaning, 1 = orthogonal, 2 = opposite.
+    #[default]
+    Cosine,
+
+    /// Euclidean (L2) distance between vectors.
+    ///
+    /// Sensitive to both direction and magnitude changes.
+    /// Useful for detecting absolute changes in embedding space.
+    Euclidean,
+
+    /// Angular distance: arccos(cosine_similarity)
+    ///
+    /// Returns the geometric angle between vectors in radians.
+    /// Range: [0, Ï€] where 0 = identical, Ï€/2 = orthogonal, Ï€ = opposite.
+    Angular,
+}
+
+/// Type alias for vector snapshot: map of NodeId to vector data
+/// Represents vector data in a snapshot, supporting both full and delta formats.
+///
+/// Full snapshots store all vectors, while delta snapshots store only changes
+/// relative to a base snapshot, significantly reducing memory usage for incremental updates.
+#[derive(Clone)]
+enum VectorSnapshot {
+    /// Full snapshot containing all vectors
+    Full(Arc<HashMap<NodeId, Arc<[f32]>>>),
+
+    /// Delta snapshot containing only changes relative to a base
+    Delta {
+        /// Timestamp of the base full snapshot
+        base_time: Timestamp,
+        /// Vectors added or updated since base
+        added: Arc<HashMap<NodeId, Arc<[f32]>>>,
+        /// Vectors removed since base
+        removed: Arc<HashSet<NodeId>>,
+    },
+}
+
+impl VectorSnapshot {
+    /// Get a vector from this snapshot, given access to the full snapshot data.
+    ///
+    /// Uses iterative traversal through delta chain to avoid stack overflow.
+    /// Enforces MAX_DELTA_CHAIN_DEPTH to prevent unbounded traversal.
+    ///
+    /// Returns:
+    /// - Ok(Some(vector)) - Vector found
+    /// - Ok(None) - Vector not found or was removed
+    /// - Err - Delta chain depth exceeded or corrupted snapshot state
+    fn get_vector(
+        &self,
+        node_id: &NodeId,
+        all_snapshots: &BTreeMap<Timestamp, VectorSnapshot>,
+    ) -> Result<Option<Arc<[f32]>>> {
+        let mut current = self;
+        let mut depth = 0;
+
+        // Iteratively traverse the delta chain with depth limit
+        loop {
+            // SAFETY: Check depth limit to prevent unbounded traversal
+            // If chain exceeds MAX_DELTA_CHAIN_DEPTH, return error instead of silently failing
+            if depth >= MAX_DELTA_CHAIN_DEPTH {
+                return Err(VectorError::IndexError(format!(
+                    "Delta chain depth exceeded {} for node {:?}. \
+                         This indicates corrupted snapshot state or misconfiguration. \
+                         Reduce full_snapshot_interval or check snapshot integrity.",
+                    MAX_DELTA_CHAIN_DEPTH, node_id
+                ))
+                .into());
+            }
+
+            match current {
+                VectorSnapshot::Full(vectors) => {
+                    return Ok(vectors.get(node_id).cloned());
+                }
+                VectorSnapshot::Delta {
+                    base_time,
+                    added,
+                    removed,
+                } => {
+                    // First check if removed
+                    if removed.contains(node_id) {
+                        return Ok(None);
+                    }
+
+                    // Then check if in added/updated
+                    if let Some(vec) = added.get(node_id) {
+                        return Ok(Some(Arc::clone(vec)));
+                    }
+
+                    // Continue to base snapshot
+                    if let Some(base) = all_snapshots.get(base_time) {
+                        current = base;
+                        depth += 1;
+                    } else {
+                        // Base snapshot was pruned - this is corrupted state
+                        return Err(VectorError::IndexError(format!(
+                            "Base snapshot at timestamp {} not found for node {:?}. \
+                                 Snapshot state is corrupted or base was incorrectly pruned.",
+                            base_time, node_id
+                        ))
+                        .into());
+                    }
+                }
+            }
+        }
+    }
+
+    /// Reconstruct all vectors in this snapshot as a HashMap.
+    ///
+    /// For delta snapshots, this combines the base with added/removed changes.
+    /// Uses iterative traversal with depth limiting to prevent stack overflow.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if:
+    /// - Delta chain depth exceeds MAX_DELTA_CHAIN_DEPTH (corrupted state)
+    /// - Base snapshot is missing (corrupted state or incorrect pruning)
+    fn to_hashmap(
+        &self,
+        all_snapshots: &BTreeMap<Timestamp, VectorSnapshot>,
+    ) -> Result<HashMap<NodeId, Arc<[f32]>>> {
+        // Use iterative approach to prevent stack overflow (similar to get_vector)
+        let mut current = self;
+        let mut depth = 0;
+
+        // Collect the chain of deltas from newest to oldest
+        struct DeltaLayer<'a> {
+            added: &'a Arc<HashMap<NodeId, Arc<[f32]>>>,
+            removed: &'a Arc<HashSet<NodeId>>,
+        }
+        let mut delta_layers: Vec<DeltaLayer<'_>> = Vec::new();
+
+        // Walk backwards through delta chain to find the Full snapshot base
+        let base_vectors: HashMap<NodeId, Arc<[f32]>> = loop {
+            if depth >= MAX_DELTA_CHAIN_DEPTH {
+                // Return error instead of partial results to prevent silent data loss
+                return Err(VectorError::IndexError(format!(
+                    "Delta chain depth exceeded {} in to_hashmap(). \
+                         This indicates corrupted snapshot state or misconfiguration. \
+                         Reduce full_snapshot_interval or check snapshot integrity.",
+                    MAX_DELTA_CHAIN_DEPTH
+                ))
+                .into());
+            }
+
+            match current {
+                VectorSnapshot::Full(vectors) => {
+                    // Found the base Full snapshot
+                    break (**vectors).clone();
+                }
+                VectorSnapshot::Delta {
+                    base_time,
+                    added,
+                    removed,
+                } => {
+                    // Record this delta layer for later application
+                    delta_layers.push(DeltaLayer { added, removed });
+
+                    // Move to base snapshot
+                    if let Some(base_snapshot) = all_snapshots.get(base_time) {
+                        current = base_snapshot;
+                        depth += 1;
+                    } else {
+                        // Base was pruned - return error to prevent silent data loss
+                        return Err(VectorError::IndexError(format!(
+                            "Base snapshot at timestamp {} not found in to_hashmap(). \
+                                 Snapshot state is corrupted or base was incorrectly pruned.",
+                            base_time
+                        ))
+                        .into());
+                    }
+                }
+            }
+        };
+
+        // Apply delta layers in reverse order (oldest to newest)
+        let mut result = base_vectors;
+        for layer in delta_layers.iter().rev() {
+            // Apply removals
+            for node_id in layer.removed.iter() {
+                result.remove(node_id);
+            }
+            // Apply additions/updates
+            for (node_id, vector) in layer.added.iter() {
+                result.insert(*node_id, Arc::clone(vector));
+            }
+        }
+
+        Ok(result)
+    }
+
+    /// Returns an approximate count of vectors in this snapshot.
+    ///
+    /// **IMPORTANT**: For delta snapshots, this returns ONLY the count of added vectors,
+    /// ignoring the base snapshot size and removed vectors. This is intentionally an
+    /// underestimate used for capacity estimation during index construction.
+    ///
+    /// **For exact counts**, use `to_hashmap().len()` which reconstructs the full snapshot
+    /// by applying all deltas to the base. Note that reconstruction has O(depth) cost and
+    /// may return partial results if the delta chain exceeds `MAX_DELTA_CHAIN_DEPTH`.
+    ///
+    /// # Examples
+    ///
+    /// ```ignore
+    /// // Delta snapshot: base has 100 vectors, added 10, removed 5
+    /// snapshot.len()           // Returns 10 (approximation, added only)
+    /// snapshot.to_hashmap().len()  // Returns 105 (exact: 100 + 10 - 5)
+    /// ```
+    fn len(&self) -> usize {
+        match self {
+            VectorSnapshot::Full(vectors) => vectors.len(),
+            VectorSnapshot::Delta { added, .. } => {
+                // Approximation: just return added size
+                // This is used only for capacity estimation during index construction
+                // and intentionally underestimates to avoid excessive memory allocation
+                added.len().max(MIN_CAPACITY_ESTIMATE)
+            }
+        }
+    }
+
+    /// Collect all vectors in this snapshot into a Vec.
+    ///
+    /// For delta snapshots, this reconstructs the full set.
+    /// Returns a vector of (NodeId, Arc<[f32]>) pairs.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the snapshot state is corrupted (see `to_hashmap` for details).
+    fn collect_all(
+        &self,
+        all_snapshots: &BTreeMap<Timestamp, VectorSnapshot>,
+    ) -> Result<Vec<(NodeId, Arc<[f32]>)>> {
+        let hashmap = self.to_hashmap(all_snapshots)?;
+        Ok(hashmap.into_iter().collect())
+    }
+}
+
+/// Storage structure for snapshot data.
+/// Can be either a full HNSW index or a delta index.
+#[derive(Clone)]
+enum SnapshotIndex {
+    Full(Arc<HnswIndex>),
+    Delta(Arc<DeltaIndex>),
+}
+
+impl std::fmt::Debug for SnapshotIndex {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            SnapshotIndex::Full(index) => f
+                .debug_struct("SnapshotIndex::Full")
+                .field("len", &index.len())
+                .field("dimensions", &index.dimensions())
+                .finish(),
+            SnapshotIndex::Delta(delta) => f
+                .debug_struct("SnapshotIndex::Delta")
+                .field("base_len", &delta.base.len())
+                .field("added_len", &delta.added.len())
+                .field("removed_len", &delta.removed.len())
+                .field(
+                    "total_len",
+                    &(delta.base.len() + delta.added.len() - delta.removed.len()),
+                )
+                .finish(),
+        }
+    }
+}
+
+impl SnapshotIndex {
+    fn search(&self, query: &[f32], k: usize) -> Result<Vec<(NodeId, f32)>> {
+        match self {
+            SnapshotIndex::Full(index) => index.search(query, k),
+            SnapshotIndex::Delta(delta) => delta.search(query, k),
+        }
+    }
+
+    fn search_with_filter(
+        &self,
+        query: &[f32],
+        k: usize,
+        predicate: &(dyn Fn(&NodeId) -> bool + Send + Sync),
+    ) -> Result<Vec<(NodeId, f32)>> {
+        match self {
+            SnapshotIndex::Full(index) => index.search_with_filter(query, k, predicate),
+            SnapshotIndex::Delta(delta) => delta.search_with_filter(query, k, predicate),
+        }
+    }
+
+    fn len(&self) -> usize {
+        match self {
+            SnapshotIndex::Full(index) => index.len(),
+            SnapshotIndex::Delta(delta) => {
+                // Correct count: base + added - removed
+                // removed.len() is O(1), so no performance penalty
+                delta.base.len() + delta.added.len() - delta.removed.len()
+            }
+        }
+    }
+
+    fn dimensions(&self) -> usize {
+        match self {
+            SnapshotIndex::Full(index) => index.dimensions(),
+            SnapshotIndex::Delta(delta) => delta.added.dimensions(),
+        }
+    }
+}
+
+/// A delta snapshot that stores only changes relative to a base snapshot.
+#[derive(Clone)]
+struct DeltaIndex {
+    /// The base snapshot this delta is built upon (usually a Full snapshot)
+    base: Arc<SnapshotIndex>,
+    /// Vectors added or updated since the base snapshot
+    added: Arc<HnswIndex>,
+    /// IDs of vectors that were removed or updated (invalidating the base version)
+    removed: Arc<HashSet<NodeId>>,
+}
+
+impl std::fmt::Debug for DeltaIndex {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("DeltaIndex")
+            .field("base", &self.base)
+            .field("added_len", &self.added.len())
+            .field("added_dimensions", &self.added.dimensions())
+            .field("removed_count", &self.removed.len())
+            .finish()
+    }
+}
+
+impl DeltaIndex {
+    fn search(&self, query: &[f32], k: usize) -> Result<Vec<(NodeId, f32)>> {
+        // Search strategy: Search for k*2 candidates from each index to ensure we don't
+        // miss true top-k when merging. This is necessary because the global top-k might
+        // be distributed across both indexes.
+        //
+        // Example: If added has [0.99, 0.95, 0.90] and base has [0.97, 0.93, 0.88],
+        // searching for k=2 in each would give [0.99, 0.95] + [0.97, 0.93].
+        // Merging gives true top-2: [0.99, 0.97] ✓
+        let search_k = k.saturating_mul(2).max(k + 10);
+
+        // 1. Search added vectors (new and updated)
+        let mut results = self.added.search(query, search_k)?;
+
+        // 2. Search base vectors with filter
+        // Filter out any ID that is in the 'removed' set, which includes:
+        // - Nodes that were updated (old version in base, new version in added)
+        // - Nodes that were removed (present in base, deleted from current state)
+        let removed = &self.removed;
+        let base_results = self
+            .base
+            .search_with_filter(query, search_k, &|id| !removed.contains(id))?;
+
+        // 3. Merge results from both indexes
+        results.extend(base_results);
+
+        // 4. Deduplicate: Although the removed filter should prevent duplicates,
+        // we deduplicate as a safety measure to ensure correctness.
+        // We keep the first occurrence (which has the better score after sorting).
+        use std::collections::HashSet;
+        let mut seen = HashSet::new();
+        results.retain(|(id, _)| seen.insert(*id));
+
+        // 5. Sort by similarity (descending) and truncate to k
+        results.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+        results.truncate(k);
+
+        Ok(results)
+    }
+
+    fn search_with_filter(
+        &self,
+        query: &[f32],
+        k: usize,
+        predicate: &(dyn Fn(&NodeId) -> bool + Send + Sync),
+    ) -> Result<Vec<(NodeId, f32)>> {
+        // Search for k*2 candidates to ensure global top-k (same strategy as search())
+        let search_k = k.saturating_mul(2).max(k + 10);
+
+        // Combine user predicate with our removed set
+        let removed = &self.removed;
+        let combined_predicate = |id: &NodeId| predicate(id) && !removed.contains(id);
+
+        // Search added (using user predicate only)
+        let mut results = self.added.search_with_filter(query, search_k, predicate)?;
+
+        // Search base (using combined predicate to filter out removed/updated nodes)
+        let base_results = self
+            .base
+            .search_with_filter(query, search_k, &combined_predicate)?;
+
+        // Merge results
+        results.extend(base_results);
+
+        // Deduplicate: Although the combined predicate should prevent duplicates,
+        // we deduplicate as a safety measure to ensure correctness.
+        // We keep the first occurrence (which has the better score after sorting).
+        use std::collections::HashSet;
+        let mut seen = HashSet::new();
+        results.retain(|(id, _)| seen.insert(*id));
+
+        // Sort and truncate to k
+        results.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+        results.truncate(k);
+
+        Ok(results)
+    }
+}
+
+/// Snapshot data protected by a single lock.
+///
+/// Groups snapshots and vector history together to ensure atomic updates
+/// and prevent deadlocks from acquiring multiple locks sequentially.
+struct SnapshotData {
+    /// Historical HNSW snapshots at anchor timestamps
+    /// Key: Timestamp when snapshot was created
+    /// Value: (Stable snapshot ID, SnapshotIndex)
+    snapshots: BTreeMap<Timestamp, (usize, SnapshotIndex)>,
+
+    /// Historical vector values at each snapshot
+    /// Key: Timestamp when snapshot was created
+    /// Value: Immutable map of NodeId -> Vector for that snapshot
+    vector_history: BTreeMap<Timestamp, VectorSnapshot>,
+}
+
+impl SnapshotData {
+    fn new() -> Self {
+        SnapshotData {
+            snapshots: BTreeMap::new(),
+            vector_history: BTreeMap::new(),
+        }
+    }
+
+    fn insert(
+        &mut self,
+        timestamp: Timestamp,
+        stable_id: usize,
+        snapshot: SnapshotIndex,
+        vectors: VectorSnapshot,
+    ) {
+        self.snapshots.insert(timestamp, (stable_id, snapshot));
+        self.vector_history.insert(timestamp, vectors);
+    }
+
+    fn remove_oldest(&mut self) {
+        if let Some(oldest_key) = self.snapshots.keys().next().copied() {
+            self.snapshots.remove(&oldest_key);
+            self.vector_history.remove(&oldest_key);
+        }
+    }
+
+    fn len(&self) -> usize {
+        self.snapshots.len()
+    }
+}
+
+/// Metadata for snapshot management.
+///
+/// Tracks state needed to determine when to create the next snapshot.
+#[derive(Debug, Clone)]
+struct SnapshotMetadata {
+    /// Last snapshot timestamp (microseconds since epoch)
+    last_snapshot_time: Timestamp,
+
+    /// Transaction count since last snapshot
+    transactions_since_snapshot: usize,
+
+    /// Vectors changed since last snapshot (resets every snapshot)
+    vectors_changed_since_snapshot: HashSet<NodeId>,
+
+    /// Total snapshots created (for ID generation)
+    total_snapshots: usize,
+
+    /// Time of the last FULL snapshot
+    last_full_snapshot_time: Timestamp,
+
+    /// Accumulated changes since the last FULL snapshot.
+    /// This is used to build Delta snapshots.
+    /// Resets only when a FULL snapshot is created.
+    ///
+    /// **MEMORY GROWTH WARNING**: This set grows unboundedly between full snapshots.
+    /// In write-heavy workloads with large `full_snapshot_interval` values, this can
+    /// consume significant memory. Mitigation strategies:
+    /// - Reduce `full_snapshot_interval` to trigger full snapshots more frequently
+    /// - Use manual snapshots (`create_manual_snapshot()`) during idle periods
+    /// - Monitor memory usage if updating >100k unique vectors between full snapshots
+    changes_accumulated: HashSet<NodeId>,
+
+    /// Number of snapshots created since the last FULL snapshot.
+    /// Used to trigger periodic FULL snapshots.
+    snapshots_since_full: usize,
+}
+
+impl SnapshotMetadata {
+    fn new(initial_time: Timestamp) -> Self {
+        SnapshotMetadata {
+            last_snapshot_time: initial_time,
+            transactions_since_snapshot: 0,
+            vectors_changed_since_snapshot: HashSet::new(),
+            total_snapshots: 0,
+            last_full_snapshot_time: initial_time,
+            changes_accumulated: HashSet::new(),
+            snapshots_since_full: 0,
+        }
+    }
+
+    /// Record a vector change for snapshot tracking.
+    fn record_change(&mut self, node_id: NodeId) {
+        self.vectors_changed_since_snapshot.insert(node_id);
+        self.changes_accumulated.insert(node_id);
+    }
+
+    /// Record a transaction (increment counter).
+    fn record_transaction(&mut self) {
+        self.transactions_since_snapshot += 1;
+    }
+
+    /// Reset tracking after creating a snapshot.
+    fn reset(&mut self, snapshot_time: Timestamp, is_full: bool) {
+        self.last_snapshot_time = snapshot_time;
+        self.transactions_since_snapshot = 0;
+        self.vectors_changed_since_snapshot.clear();
+        self.total_snapshots += 1;
+
+        if is_full {
+            self.last_full_snapshot_time = snapshot_time;
+            self.changes_accumulated.clear();
+            self.snapshots_since_full = 0;
+        } else {
+            self.snapshots_since_full += 1;
+        }
+    }
+}
+
+/// Combined state for current vectors and metadata.
+///
+/// **Issue #233 Optimization**: This struct combines vector storage and metadata
+/// into a single structure protected by one RwLock, reducing lock acquisitions
+/// from 3 to 1 per add() operation.
+///
+/// This eliminates the need for:
+/// - DashMap internal locking for vectors
+/// - Separate RwLock for metadata
+///
+/// Instead, we acquire a single write lock to update both vectors and metadata atomically.
+#[derive(Debug)]
+struct VectorState {
+    /// Current vector storage - maintains actual vector data for snapshot copying
+    /// Maps NodeId to the vector embedding
+    vectors: HashMap<NodeId, Arc<[f32]>>,
+
+    /// Metadata for snapshot management
+    metadata: SnapshotMetadata,
+}
+
+impl VectorState {
+    fn new(initial_time: Timestamp) -> Self {
+        VectorState {
+            vectors: HashMap::new(),
+            metadata: SnapshotMetadata::new(initial_time),
+        }
+    }
+}
+
+/// Temporal vector index for time-aware semantic search.
+///
+/// Maintains a current HNSW index plus historical snapshots at configurable intervals,
+/// enabling queries like "find similar vectors as of timestamp T" and semantic drift
+/// tracking.
+///
 /// # Thread Safety
 ///
 /// This struct is thread-safe using `RwLock` for internal mutability. Multiple threads
@@ -1335,10 +2173,185 @@ impl TemporalVectorIndex {
         self.current.distance_metric()
     }
 }
+
+/// Snapshot information for monitoring.
+#[derive(Debug, Clone)]
+pub struct SnapshotInfo {
+    /// Sequential snapshot ID
+    pub snapshot_id: usize,
+    /// Timestamp when snapshot was created (microseconds since epoch)
+    pub timestamp: Timestamp,
+    /// Number of vectors in snapshot
+    pub vector_count: usize,
+    /// Estimated size in bytes
+    pub size_bytes: usize,
+    /// Age of snapshot
+    pub age: Duration,
+}
+
+/// Observer adapter for TemporalVectorIndex to react to storage events.
+///
+/// This struct implements `StorageObserver` to enable TemporalVectorIndex to create
+/// snapshots when HistoricalStorage creates anchors, maintaining synchronization
+/// between graph versioning and vector indexing.
+///
+/// # Design
+///
+/// The observer wraps an Arc reference to TemporalVectorIndex, allowing multiple
+/// components to share the same index. When node/edge anchors are created, the
+/// observer triggers snapshot creation and stores the snapshot ID in the anchor's
+/// metadata via the return value.
+///
+/// # Example
+///
+/// ```no_run
+/// use gallifreydb::index::vector::temporal::{TemporalVectorIndex, VectorIndexObserver, TemporalVectorConfig};
+/// use gallifreydb::index::vector::{HnswConfig, DistanceMetric};
+/// use gallifreydb::storage::historical::HistoricalStorage;
+/// use std::sync::Arc;
+///
+/// # fn example() -> gallifreydb::utils::Result<()> {
+/// // Create temporal vector index
+/// let config = TemporalVectorConfig::default_with_hnsw(
+///     HnswConfig::new(384, DistanceMetric::Cosine)
+/// );
+/// let index = Arc::new(TemporalVectorIndex::new(config)?);
+///
+/// // Create observer wrapper
+/// let observer = VectorIndexObserver::new(Arc::clone(&index));
+///
+/// // Register with HistoricalStorage
+/// let mut storage = HistoricalStorage::new();
+/// storage.add_observer(Arc::new(observer));
+///
+/// // Now anchors will automatically trigger vector snapshots
+/// # Ok(())
+/// # }
+/// ```
+pub struct VectorIndexObserver {
+    /// The temporal vector index (reserved for future use in metrics/monitoring)
+    #[allow(dead_code)]
+    index: Arc<TemporalVectorIndex>,
+}
+
+impl VectorIndexObserver {
+    /// Creates a new observer for a TemporalVectorIndex.
+    ///
+    /// # Arguments
+    ///
+    /// * `index` - Arc reference to the TemporalVectorIndex to observe
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// # use gallifreydb::index::vector::temporal::{TemporalVectorIndex, VectorIndexObserver, TemporalVectorConfig};
+    /// # use gallifreydb::index::vector::{HnswConfig, DistanceMetric};
+    /// # use std::sync::Arc;
+    /// # fn example() -> gallifreydb::utils::Result<()> {
+    /// let config = TemporalVectorConfig::default_with_hnsw(
+    ///     HnswConfig::new(384, DistanceMetric::Cosine)
+    /// );
+    /// let index = Arc::new(TemporalVectorIndex::new(config)?);
+    /// let observer = VectorIndexObserver::new(index);
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn new(index: Arc<TemporalVectorIndex>) -> Self {
+        VectorIndexObserver { index }
+    }
+}
+
+impl crate::core::observer::StorageObserver for VectorIndexObserver {
+    fn on_event(&self, event: &crate::core::observer::StorageEvent) -> crate::utils::Result<()> {
+        use crate::core::observer::StorageEvent;
+
+        match event {
+            // Note: The PreAnchorHook already creates the snapshot and links the snapshot_id
+            // to the anchor. This observer is only for post-commit actions like logging/metrics.
+            StorageEvent::NodeAnchorCreated {
+                node_id: _node_id,
+                timestamp: _timestamp,
+                ..
+            } => {
+                #[cfg(feature = "observability")]
+                tracing::debug!(
+                    "VectorIndexObserver: Node anchor created for {} at timestamp {} (snapshot already created by pre-anchor hook)",
+                    _node_id,
+                    _timestamp
+                );
+                Ok(())
+            }
+            StorageEvent::EdgeAnchorCreated {
+                edge_id: _edge_id,
+                timestamp: _timestamp,
+                ..
+            } => {
+                #[cfg(feature = "observability")]
+                tracing::debug!(
+                    "VectorIndexObserver: Edge anchor created for {} at timestamp {} (snapshot already created by pre-anchor hook)",
+                    _edge_id,
+                    _timestamp
+                );
+                Ok(())
+            }
+            // Ignore other events
+            _ => Ok(()),
+        }
+    }
+
+    fn interested_in(&self, event: &crate::core::observer::StorageEvent) -> bool {
+        // Only interested in anchor creation events for logging/metrics
+        event.is_anchor_event()
+    }
+}
+
+/// Memory usage statistics for monitoring.
+///
+/// Use these metrics to detect potential memory growth and optimize snapshot configuration.
+#[derive(Debug, Clone)]
+pub struct MemoryStats {
+    /// Number of unique vectors changed since last FULL snapshot.
+    ///
+    /// **WARNING**: This grows unboundedly between full snapshots.
+    /// If this exceeds 100k, consider reducing `full_snapshot_interval`
+    /// or calling `create_manual_snapshot()`.
+    pub changes_accumulated_size: usize,
+
+    /// Number of vectors changed since last snapshot (any type).
+    ///
+    /// Resets after each snapshot creation.
+    pub vectors_changed_since_snapshot: usize,
+
+    /// Number of snapshots created since last full snapshot.
+    pub snapshots_since_full: usize,
+
+    /// Total number of snapshots stored.
+    pub total_snapshots: usize,
+
+    /// Number of vectors in current (live) index.
+    pub current_vectors: usize,
+}
+
+impl MemoryStats {
+    /// Checks if memory growth is concerning and may require intervention.
+    ///
+    /// Returns true if `changes_accumulated_size` exceeds 100,000 entries,
+    /// which could indicate excessive memory usage.
+    pub fn is_high_memory_usage(&self) -> bool {
+        self.changes_accumulated_size > 100_000
+    }
+
+    /// Estimates memory overhead from accumulated changes in bytes.
+    ///
+    /// This is a rough estimate assuming ~8 bytes per NodeId in the HashSet.
+    pub fn estimated_accumulated_bytes(&self) -> usize {
+        self.changes_accumulated_size * 8 // NodeId is u64
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::index::vector::HnswConfig;
 
     fn create_test_index() -> Result<TemporalVectorIndex> {
         let config = TemporalVectorConfig {
@@ -3388,9 +4401,7 @@ mod tests {
         if let Err(err) = result {
             let err_msg = err.to_string();
             assert!(
-                err_msg.contains("not found")
-                    || err_msg.contains("corrupted")
-                    || err_msg.contains("missing"),
+                err_msg.contains("not found") || err_msg.contains("corrupted"),
                 "Error should mention missing base or corrupted state, got: {}",
                 err_msg
             );
@@ -3406,9 +4417,7 @@ mod tests {
         if let Err(err) = result {
             let err_msg = err.to_string();
             assert!(
-                err_msg.contains("not found")
-                    || err_msg.contains("corrupted")
-                    || err_msg.contains("missing"),
+                err_msg.contains("not found") || err_msg.contains("corrupted"),
                 "Error should mention missing base or corrupted state, got: {}",
                 err_msg
             );
@@ -3659,5 +4668,3 @@ mod tests {
         Ok(())
     }
 }
-#[cfg(test)]
-mod coverage_tests;
